Fixed handling of legacy otp for 2FA
